+++ benchmarks/burgers/burgers_package.cpp
@@ -18,12 +18,12 @@
 #include <string>
 #include <vector>
 
+#include "coordinates/coordinates.hpp"
 #include "defs.hpp"
 #include "kokkos_abstraction.hpp"
+#include "parthenon/package.hpp"
 #include "reconstruct/dc_inline.hpp"
 #include "utils/error_checking.hpp"
-#include <coordinates/coordinates.hpp>
-#include <parthenon/package.hpp>
 
 #include "burgers_package.hpp"
 #include "recon.hpp"