+++ src/interface/variable.hpp
@@ -39,6 +39,7 @@
 #include "defs.hpp"
 #include "interface/metadata.hpp"
 #include "interface/variable_state.hpp"
+#include "mesh/refinement_in_one.hpp"
 #include "parthenon_arrays.hpp"
 #include "utils/error_checking.hpp"
 
@@ -99,6 +100,12 @@ class CellVariable {
 
   ///< retrieve metadata for variable
   inline Metadata metadata() const { return m_; }
+
+  /// Refinement functions owned in metadata
+  inline bool IsRefined() const { return m_.IsRefined(); }
+  inline const refinement::RefinementFunctions_t &GetRefinementFunctions() const {
+    return m_.GetRefinementFunctions();
+  }
 
   /// Get Sparse ID (InvalidSparseID if not sparse)
   inline int GetSparseID() const { return IsSparse() ? sparse_id_ : InvalidSparseID; }