+++ src/interface/variable_state.hpp
@@ -16,19 +16,19 @@
 #include <limits>
 
 #include "basic_types.hpp"
+#include "defs.hpp"
 #include "globals.hpp"
-#include "interface/metadata.hpp"
 #include "parthenon_arrays.hpp"
 
 namespace parthenon {
 
+// forward declaration
+class Metadata;
+
 static constexpr int InvalidSparseID = std::numeric_limits<int>::min();
 
 struct VariableState : public empty_state_t {
-  explicit VariableState(const Metadata &md, int sparse_id = InvalidSparseID)
-      : allocation_threshold(md.GetAllocationThreshold()),
-        deallocation_threshold(md.GetDeallocationThreshold()),
-        sparse_default_val(md.GetDefaultValue()), sparse_id(sparse_id) {}
+  explicit VariableState(const Metadata &md, int sparse_id = InvalidSparseID);
 
   KOKKOS_INLINE_FUNCTION
   VariableState(Real alloc, Real dealloc, Real sparse_default_val = 0.0,