+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Changed (changing behavior/API/variables/...)
 - [[PR 753]](https://github.com/parthenon-hpc-lab/parthenon/pull/753) Cleanup uniform Cartesian variable names
+- [[PR 769]](https://github.com/parthenon-hpc-lab/parthenon/pull/769) Thread custom prolongation-restriction functions through infrastructure and to userspace
 - [[PR 781]](https://github.com/lanl/parthenon/pull/781) Output variables with same shape as in parthenon
 - [[PR 758]](https://github.com/lanl/parthenon/pull/758) Bump required C++ standard to C++17
 - [[PR 710]](https://github.com/lanl/parthenon/pull/710) Remove data transpose in hdf5 and restart outputs